Split code between library and binary
